Split taiko extended rpc to another class
