+++ src/Nethermind/Nethermind.Taiko/Rpc/TaikoExtendedEthModule.cs
@@ -1,0 +1,42 @@
+// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+using System.Threading.Tasks;
+using Nethermind.Blockchain.Synchronization;
+using Nethermind.Int256;
+using Nethermind.JsonRpc;
+
+namespace Nethermind.Taiko.Rpc;
+
+public class TaikoExtendedEthModule(
+    ISyncConfig syncConfig,
+    IL1OriginStore l1OriginStore) : ITaikoExtendedEthRpcModule
+{
+    private static readonly ResultWrapper<L1Origin?> NotFound = ResultWrapper<L1Origin?>.Fail("not found");
+
+    public Task<ResultWrapper<string>> taiko_getSyncMode() => ResultWrapper<string>.Success(syncConfig switch
+    {
+        { SnapSync: true } => "snap",
+        _ => "full",
+    });
+
+    public Task<ResultWrapper<L1Origin?>> taiko_headL1Origin()
+    {
+        UInt256? head = l1OriginStore.ReadHeadL1Origin();
+        if (head is null)
+        {
+            return NotFound;
+        }
+
+        L1Origin? origin = l1OriginStore.ReadL1Origin(head.Value);
+
+        return origin is null ? NotFound : ResultWrapper<L1Origin?>.Success(origin);
+    }
+
+    public Task<ResultWrapper<L1Origin?>> taiko_l1OriginByID(UInt256 blockId)
+    {
+        L1Origin? origin = l1OriginStore.ReadL1Origin(blockId);
+
+        return origin is null ? NotFound : ResultWrapper<L1Origin?>.Success(origin);
+    }
+}
