+++ src/Nethermind/Nethermind.Taiko/TaikoNethermindApi.cs
@@ -1,11 +1,12 @@
 // SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
+using Autofac;
 using Nethermind.Api;
 
 namespace Nethermind.Taiko;
 
 public class TaikoNethermindApi(NethermindApi.Dependencies dependencies) : NethermindApi(dependencies)
 {
-    public L1OriginStore? L1OriginStore { get; set; }
+    public IL1OriginStore L1OriginStore => Context.Resolve<IL1OriginStore>();
 }
