+++ src/Nethermind/Nethermind.Taiko/TaikoPlugin.cs
@@ -47,7 +47,6 @@ namespace Nethermind.Taiko;
 public class TaikoPlugin(ChainSpec chainSpec) : IConsensusPlugin
 {
     public const string Taiko = "Taiko";
-    private const string L1OriginDbName = "L1Origin";
     public string Author => "Nethermind";
     public string Name => Taiko;
     public string Description => "Taiko support for Nethermind";
@@ -56,7 +55,6 @@ public class TaikoPlugin(ChainSpec chainSpec) : IConsensusPlugin
     private ILogger _logger;
 
     private IMergeConfig _mergeConfig = null!;
-    private ISyncConfig _syncConfig = null!;
 
     private IBlockCacheService? _blockCacheService;
 
@@ -66,7 +64,6 @@ public class TaikoPlugin(ChainSpec chainSpec) : IConsensusPlugin
     {
         _api = (TaikoNethermindApi)api;
         _mergeConfig = _api.Config<IMergeConfig>();
-        _syncConfig = _api.Config<ISyncConfig>();
         _logger = _api.LogManager.GetClassLogger();
 
         ArgumentNullException.ThrowIfNull(_api.BlockTree);
@@ -80,22 +77,6 @@ public class TaikoPlugin(ChainSpec chainSpec) : IConsensusPlugin
         _api.BlockPreprocessor.AddFirst(new MergeProcessingRecoveryStep(_api.Context.Resolve<IPoSSwitcher>()));
 
         return Task.CompletedTask;
-    }
-
-    public void InitTxTypesAndRlpDecoders(INethermindApi api)
-    {
-        _api = (TaikoNethermindApi)api;
-
-        ArgumentNullException.ThrowIfNull(_api.DbProvider);
-        ArgumentNullException.ThrowIfNull(_api.DbFactory);
-
-        IRlpStreamDecoder<L1Origin> r1OriginDecoder = new L1OriginDecoder();
-        Rlp.RegisterDecoder(typeof(L1Origin), r1OriginDecoder);
-
-
-        IDb db = _api.DbFactory.CreateDb(new DbSettings(L1OriginDbName, L1OriginDbName.ToLower()));
-        _api.DbProvider!.RegisterDb(L1OriginDbName, db);
-        _api.L1OriginStore = new(_api.DbProvider.GetDb<IDb>(L1OriginDbName), r1OriginDecoder);
     }
 
     public Task InitRpcModules()
@@ -194,7 +175,6 @@ public class TaikoPlugin(ChainSpec chainSpec) : IConsensusPlugin
 
     private static TaikoPayloadPreparationService CreatePayloadPreparationService(TaikoNethermindApi api, IRlpStreamDecoder<Transaction> txDecoder)
     {
-        ArgumentNullException.ThrowIfNull(api.L1OriginStore);
         ArgumentNullException.ThrowIfNull(api.SpecProvider);
 
         IReadOnlyTxProcessorSource txProcessingEnv = api.ReadOnlyTxProcessingEnvFactory.Create();
@@ -280,11 +260,18 @@ public class TaikoModule : Module
             .Map<TaikoChainSpecEngineParameters, ChainSpec>(chainSpec =>
                 chainSpec.EngineChainSpecParametersProvider.GetChainSpecParameters<TaikoChainSpecEngineParameters>())
 
+            // L1 origin store
+            .AddSingleton<IRlpStreamDecoder<L1Origin>, L1OriginDecoder>()
+            .AddKeyedSingleton<IDb>(L1OriginStore.L1OriginDbName, (ctx) => ctx
+                .Resolve<IDbFactory>().CreateDb(new DbSettings(L1OriginStore.L1OriginDbName, L1OriginStore.L1OriginDbName.ToLower())))
+            .AddSingleton<IL1OriginStore, L1OriginStore>()
+
+            // Sync modification
             .AddSingleton<IPoSSwitcher>(AlwaysPoS.Instance)
             .AddSingleton<StartingSyncPivotUpdater, UnsafeStartingSyncPivotUpdater>()
-
             .AddDecorator<BeaconSync>((_, strategy) =>
             {
+                // Normally not turned on at start because `StartingSyncPivotUpdater` waiting for pivot
                 strategy.AllowBeaconHeaderSync();
                 return strategy;
             })
@@ -298,6 +285,15 @@ public class TaikoModule : Module
 
             .AddSingleton<IHealthHintService, IBlocksConfig>((blocksConfig) =>
                 new ManualHealthHintService(blocksConfig.SecondsPerSlot * 6, HealthHintConstants.InfinityHint))
+
+            // Rpc
+            .RegisterSingletonJsonRpcModule<ITaikoExtendedEthRpcModule, TaikoExtendedEthModule>()
+
+            // Need to set the rlp globally
+            .OnBuild((ctx) =>
+            {
+                Rlp.RegisterDecoder(typeof(L1Origin), ctx.Resolve<IRlpStreamDecoder<L1Origin>>());
+            })
             ;
     }
 }
